moved delete and create endpoints to the routers with the same name;
