+++ src/climbingAPI/app/routers/pitch.py
@@ -11,10 +11,33 @@ from app.core.auth import auth
 
 from app.models.pitch.pitch_model import PitchModel
 from app.models.pitch.update_pitch_model import UpdatePitchModel
-from app.models.ascent.ascent_model import AscentModel
-from app.models.ascent.create_ascent_model import CreateAscentModel
+from app.models.route.route_model import RouteModel
+from app.models.pitch.create_pitch_model import CreatePitchModel
 
 router = APIRouter()
+
+
+@router.post('/route/{route_id}', description="Create a new pitch", response_model=PitchModel, dependencies=[Depends(auth.implicit_scheme)])
+async def create_pitch(route_id: str, pitch: CreatePitchModel = Body(...), user: Auth0User = Security(auth.get_user, scopes=["write:diary"])):
+    pitch = jsonable_encoder(pitch)
+    pitch["user_id"] = user.id
+    pitch["ascent_ids"] = []
+    pitch["media_ids"] = []
+    db = await get_db()
+    new_pitch = await db["pitch"].insert_one(pitch)
+    # created pitch
+    created_pitch = await db["pitch"].find_one({"_id": new_pitch.inserted_id})
+    if created_pitch is None:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail=f"Pitch {new_pitch.inserted_id} not found")
+    # pitch was found
+    update_result = await db["route"].update_one({"_id": ObjectId(route_id)}, {"$push": {"pitch_ids": new_pitch.inserted_id}})
+    if update_result.modified_count != 1:
+        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
+                            detail=f"Pitch {new_pitch.inserted_id} was not added to route {route_id}")
+    # pitch_id was added to route
+    return JSONResponse(status_code=status.HTTP_201_CREATED,
+                        content=jsonable_encoder(PitchModel(**created_pitch)))
 
 
 @router.get('', description="Retrieve all pitches", response_model=List[PitchModel], dependencies=[Depends(auth.implicit_scheme)])
@@ -55,70 +78,23 @@ async def update_pitch(pitch_id: str, pitch: UpdatePitchModel = Body(...), user:
                         detail=f"Pitch {pitch_id} not found")
 
 
-@router.post('/{pitch_id}', description="Create a new ascent", response_model=AscentModel, dependencies=[Depends(auth.implicit_scheme)])
-async def create_ascent(pitch_id: str, ascent: CreateAscentModel = Body(...), user: Auth0User = Security(auth.get_user, scopes=["write:diary"])):
-    ascent = jsonable_encoder(ascent)
-    ascent["user_id"] = user.id
-    ascent["media_ids"] = []
+@router.delete('/{pitch_id}/route/{route_id}', description="Delete a pitch", response_model=RouteModel, dependencies=[Depends(auth.implicit_scheme)])
+async def delete_pitch(route_id: str, pitch_id: str, user: Auth0User = Security(auth.get_user, scopes=["write:diary"])):
     db = await get_db()
-    new_ascent = await db["ascent"].insert_one(ascent)
-    # created ascent
-    created_ascent = await db["ascent"].find_one({"_id": new_ascent.inserted_id})
-    if created_ascent is None:
+    route = await db["route"].find_one({"_id": ObjectId(route_id), "user_id": user.id})
+    if route is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
-                            detail=f"Ascent {new_ascent.inserted_id} not found")
-    # ascent was found
-    update_result = await db["pitch"].update_one({"_id": ObjectId(pitch_id)}, {"$push": {"ascent_ids": new_ascent.inserted_id}})
-    if update_result.modified_count != 1:
-        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                            detail=f"Ascent {new_ascent.inserted_id} was not added to pitch {pitch_id}")
-    # ascent_id was added to pitch
-    return JSONResponse(status_code=status.HTTP_201_CREATED,
-                        content=jsonable_encoder(AscentModel(**created_ascent)))
-
-
-@router.delete('/{pitch_id}/ascent/{ascent_id}', description="Delete an ascent", response_model=PitchModel, dependencies=[Depends(auth.implicit_scheme)])
-async def delete_ascent(pitch_id: str, ascent_id: str, user: Auth0User = Security(auth.get_user, scopes=["write:diary"])):
-    db = await get_db()
+                            detail=f"Route {route_id} not found")
+    # route was found
     pitch = await db["pitch"].find_one({"_id": ObjectId(pitch_id), "user_id": user.id})
     if pitch is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                             detail=f"Pitch {pitch_id} not found")
     # pitch was found
-    ascent = await db["ascent"].find_one({"_id": ObjectId(ascent_id), "user_id": user.id})
-    if ascent is None:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
-                            detail=f"Ascent {ascent_id} not found")
-    # ascent was found
-    if await db["pitch"].find_one({"_id": ObjectId(pitch_id), "ascent_ids": ObjectId(ascent_id)}) is None:
+    if await db["route"].find_one({"_id": ObjectId(route_id), "pitch_ids": ObjectId(pitch_id)}) is None:
         raise HTTPException(status_code=status.HTTP_409_CONFLICT,
-                            detail=f"Ascent {ascent_id} does not belong to pitch {pitch_id}")
-    # ascent belongs to pitch
-    delete_result = await db["ascent"].delete_one({"_id": ObjectId(ascent_id)})
-    if delete_result.deleted_count != 1:
-        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                            detail=f"Ascent {ascent_id} could not be deleted")
-    # ascent was deleted
-    update_result = await db["pitch"].update_one({"_id": ObjectId(pitch_id)}, {"$pull": {"ascent_ids": ObjectId(ascent_id)}})
-    if update_result.modified_count != 1:
-        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                            detail=f"Removing ascent_id {ascent_id} from pitch {pitch_id} failed")
-    # ascent_id was removed
-    if (updated_pitch := await db["pitch"].find_one({"_id": ObjectId(pitch_id)})) is not None:
-        return updated_pitch
-    else:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
-                            detail=f"Pitch {pitch_id} not found")
-
-
-@router.delete('/{pitch_id}', description="Delete a pitch", dependencies=[Depends(auth.implicit_scheme)])
-async def delete_pitch(pitch_id: str, user: Auth0User = Security(auth.get_user, scopes=["write:diary"])):
-    db = await get_db()
-    pitch = await db["pitch"].find_one({"_id": ObjectId(pitch_id), "user_id": user.id})
-    if pitch is None:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
-                            detail=f"Pitch {pitch_id} not found")
-    # pitch was found
+                            detail=f"Pitch {pitch_id} does not belong to route {route_id}")
+    # pitch belongs to route
     delete_result = await db["ascent"].delete_many({"_id": {"$in": pitch["ascent_ids"]}})
     if not delete_result.acknowledged:
         raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
@@ -129,4 +105,13 @@ async def delete_pitch(pitch_id: str, user: Auth0User = Security(auth.get_user, 
         raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
                             detail=f"Pitch {pitch_id} could not be deleted")
     # pitch was deleted
-    return JSONResponse(status_code=status.HTTP_204_NO_CONTENT)
+    update_result = await db["route"].update_one({"_id": ObjectId(route_id)}, {"$pull": {"pitch_ids": ObjectId(pitch_id)}})
+    if update_result.modified_count != 1:
+        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
+                            detail=f"Removing pitch_id {pitch_id} from route {route_id} failed")
+    # pitch_id was removed
+    if (updated_route := await db["route"].find_one({"_id": ObjectId(route_id)})) is not None:
+        return updated_route
+    else:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail=f"Route {route_id} not found")
